feat(本地打包配置): 新增webpack多页面打包,wp脚手架搭建,编译优化
新增webpack多页面打包,wp脚手架搭建,编译优化

+++ build/build.js
@@ -4,35 +4,39 @@
  * 生产环境
  */
 
-const rimraf = require('rimraf')
-const ora = require('ora')
-const chalk = require('chalk')
-const path = require('path')
+ module.exports = options => {
+    const rimraf = require('rimraf')
+    const ora = require('ora')
+    const chalk = require('chalk')
+    const path = require('path')
+    
+    // 删除之前编译的dist目录
+    options.clear && rimraf.sync(path.join(process.cwd(), 'dist'))
+    
+    const config = require('./base')(options)
+    const webpack = require('webpack')
+    const spinner = ora('项目开始构建中    (￣▽￣)~*' + '\n')
+    spinner.start()
+    
+    webpack(config.toConfig(), (err, stats) => {
+        spinner.stop()
+        if(err) throw err
+        process.stdout.write(stats.toString({
+            colors: true,
+            modules: false,
+            children: false,
+            chunks: false,
+            chunkModules: false,
+        }) + "\n\n")
+    
+        if(stats.hasErrors()) {
+            console.log(chalk.red('构建失败\n ┭┮﹏┭┮' + '\n'))
+            process.exit(1)
+        }
+    
+        console.log(chalk.cyan('production has been builded  (￣▽￣)~*' + '\n'))
+    })
+ }
 
-// 删除之前编译的dist目录
-rimraf.sync(path.join(process.cwd(), 'dist'))
 
-const config = require('./base')()
-const webpack = require('webpack')
-const spinner = ora('项目开始构建中    (￣▽￣)~*' + '\n')
-spinner.start()
-
-webpack(config.toConfig(), (err, stats) => {
-    spinner.stop()
-    if(err) throw err
-    process.stdout.write(stats.toString({
-        colors: true,
-        modules: false,
-        children: false,
-        chunks: false,
-        chunkModules: false,
-    }) + "\n\n")
-
-    if(stats.hasErrors()) {
-        console.log(chalk.red('构建失败\n ┭┮﹏┭┮' + '\n'))
-        process.exit(1)
-    }
-
-    console.log(chalk.cyan('production has been builded  (￣▽￣)~*' + '\n'))
-})
